Can now mark folders as untrusted
Getting some strange behaviour with nautilus though, where you may
have to first left-click the file to get it to register whether or not
the file is trusted or untrusted, before then getting the correct
behaviour (check-mark or no check-mark) when right-clicking. I assume
it may just be an issue of not resetting a variable in python, rather
than Nautilus' fault.